Handle recursive schemas correctly
This change checks for cycles when resolving schemas by keeping a cache
of each fully built schema object in a given example. If an item is in
the cache but not yet completely built, it indicates a cycle, and the
example generator bails out.

Sometimes, the recursion isn't actually necessary (e.g., for a property
that isn't required), in which case we try to intelligently omit
recursive schema references.

This should resolve #45, although without a link to the spec they're
using it's difficult to be 100% sure.