Allow importing apisprout from external packages
This can be useful for testing go code against the mock server.
The mock server could be started in a go test and the tests are not
dependant on a docker daemon or similar runtime.

To start a server or get a http.Handler without passing a file path or
URL, the server is separated into a server (`OpenAPIServer`) that only
requires an `openapi3.T` to run and the loading from a file or URL is
handled by a `ConfigReloader`. It can load from a URL or file path and
reload the `OpenAPIServer` on file changes.

Signed-off-by: leonnicolas <leonloechner@gmx.de>